requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
cloudinary==1.38.0
Django==4.2
django-ckeditor==6.7.0
django-cloudinary-storage==0.3.0
django-js-asset==2.2.0
djangorestframework==3.15.2
gunicorn==21.2.0
idna==3.7
packaging==24.0
pillow==10.3.0
python-decouple==3.8
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
